Final integration of various pieces: - registrar invocation done from generated code, removed from startup listener - we now try to correct module references at compile time (mapping from configuration   references to the modules responsible). This currently uses the naming scheme right   now which breaks when your initializers aren't the expected name - and this breaks   the tests that depend on ContextAutoConfiguration and PropertyPlaceholderAutoConfiguration   because they are handled at the moment by the custom module written in the library.   So tests changed to depend on the module for now, this would go away when the real   boot is in place (the naming scheme would be correct when looking up the initializer).   Some tests seemed to have been switched before now, and some weren't - now they have   all been switched. - With the new getRoot() on the Module interface there is no need for applications to   Import their own module, and indeed that doesn't work reliably when using APT since   you are referring to something that isn't there, or might be there on an incremental   compilation (these differences cause us to do differing things in APT).  More reliable   to not import it directly but get it implicitly included. - moved enabler/registrar tests from fails into mainline, gives 4 more passing tests.
Tests pass when run through maven or through eclipse. Some still some incremental
compilation quirks in eclipse...
