--- modules/fails/src/main/java/lib/enable/Bar.java
@@ -1,14 +1,0 @@
-package lib.enable;
-
-public class Bar {
-
-    private final Foo foo;
-
-    public Bar(Foo foo) {
-        this.foo = foo;
-    }
-
-    public Foo getFoo() {
-        return this.foo;
-    }
-}
