--- modules/fails/src/main/java/lib/enable/BarConfiguration.java
@@ -1,33 +1,0 @@
-package lib.enable;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class BarConfiguration {
-
-	@Value("${app.value}")
-	private String message;
-
-	@Bean
-	public Foo foo() {
-		return new Foo();
-	}
-
-	@Bean
-	public Bar bar(Foo foo) {
-		return new Bar(foo);
-	}
-
-	@Bean
-	public CommandLineRunner runner(Bar bar) {
-		return args -> {
-			System.out.println("Message: " + message);
-			System.out.println("Bar: " + bar);
-			System.out.println("Foo: " + bar.getFoo());
-		};
-	}
-
-}
