+++ modules/tests/src/main/java/app/importer/SampleApplication.java
@@ -1,15 +1,20 @@
 package app.importer;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.context.annotation.Import;
 
 import boot.autoconfigure.context.ContextAutoConfigurationModule;
-import lib.importer.SomeConfigurationModule;
+import lib.imported.Bar;
+import lib.importer.SomeConfiguration;
 
 @SpringBootConfiguration
-@Import({ SomeConfigurationModule.class, ContextAutoConfigurationModule.class })
+@Import({ SomeConfiguration.class, ContextAutoConfigurationModule.class })
 public class SampleApplication {
+	
+	@Autowired
+	Bar bar;
 
 	public static void main(String[] args) {
 		SpringApplication app = new SpringApplication(SampleApplication.class);
