+++ modules/tests/src/main/java/app/provider/list/SampleApplication.java
@@ -2,14 +2,12 @@ package app.provider.list;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.context.ConfigurationPropertiesAutoConfiguration;
-import org.springframework.boot.autoconfigure.context.PropertyPlaceholderAutoConfiguration;
 import org.springframework.context.annotation.Import;
 
+import boot.autoconfigure.context.ContextAutoConfigurationModule;
+
 @SpringBootConfiguration
-@Import({ SampleApplicationModule.class, SampleConfiguration.class,
-		PropertyPlaceholderAutoConfiguration.class,
-		ConfigurationPropertiesAutoConfiguration.class })
+@Import({ SampleConfiguration.class,ContextAutoConfigurationModule.class})
 public class SampleApplication {
 
 	public static void main(String[] args) {
