+++ modules/tests/src/main/java/app/statc/SampleApplication.java
@@ -4,16 +4,14 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.context.ConfigurationPropertiesAutoConfiguration;
-import org.springframework.boot.autoconfigure.context.PropertyPlaceholderAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import boot.autoconfigure.context.ContextAutoConfigurationModule;
+
 @SpringBootConfiguration
-@Import({ SampleApplicationModule.class, SampleConfiguration.class,
-		PropertyPlaceholderAutoConfiguration.class,
-		ConfigurationPropertiesAutoConfiguration.class })
+@Import({ SampleConfiguration.class,ContextAutoConfigurationModule.class})
 public class SampleApplication {
 
 	public static void main(String[] args) {
