+++ modules/tests/src/main/java/lib/enabler/SomeConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2016-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,25 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package lib.enabler;
 
-package lib.enable;
+import org.springframework.context.annotation.Configuration;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.context.annotation.Import;
+import lib.enable.EnableBar;
 
 /**
  * @author Dave Syer
  *
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Documented
-@Import(BarConfiguration.class)
-public @interface EnableBar {
+@Configuration
+@EnableBar
+public class SomeConfiguration {
 
 }
